feat: Add HourlyForecast model for weather data
This commit introduces the `HourlyForecast` struct, conforming to `Codable` and `Identifiable`.

The model includes properties for:
- `dt`: Timestamp (Int) representing the forecast time.
- `temp`: Temperature (Double) at the forecast time.
- `weather`: An array of `WeatherCondition` objects describing weather conditions.

A computed property `id` is added for `Identifiable` conformance, using the timestamp (`dt`).  A computed property `date` is added to return a `Date` object from the timestamp.

This model will be used to represent individual hourly forecast data points retrieved from the weather API.