feat: Implement SearchCityUseCase for city searching
This commit introduces the `SearchCityUseCase` and `SearchCityUseCaseProtocol`.
The `SearchCityUseCase` class implements the `SearchCityUseCaseProtocol`
and encapsulates the logic for searching cities. It depends on
`WeatherRepositoryProtocol` to fetch city data based on the provided query.
The `execute` function within the use case calls the `searchCity` function
of the `weatherRepository` passing the query and completion handler,
decoupling the UI from the data layer.